Use zIndex instead of display to hide/show tabs
This improves the performance of tab__content rendering, especially
for tabs with a lot of children, like EmojiKeyboard.

For example, if using display to control visibility, the emoji
characters won't start rendering until the tab is active, so clicking
on the emoji tab fetches the emoji.json and then renders the characters.
Yuck. Too slow. This starts that process as soon as the Tabs component
is rendered.