Use images for emoji instead of Unicode
Using characters to represent emoji was falling short:

1. I couldn't count on all characters being supported on various
OSes.
2. I needed two different functions for drawing "nipples" on the
Canvas. One for the nipple images, another for the emoji text

This change adds .png files for the latest set of Apple emoji, and
and draws emoji on the Canvas as an image instead of text. This
allowed me to merge the functions that render "nipples" on the
Canvas, and makes resizing "nipples" much simpler.